rename model directory
