+++ src/app/components/plus-minus-counter/plus-minus-counter.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-import { Counter } from '../../model/counter';
+import { Counter } from '../../models/counter';
 import { CounterService } from '../../services/counter.service';
 
 @Component({
